# test: MySql test performance improvements