--- a/tests/Agent/IntegrationTests/IntegrationTestHelpers/IntegrationTestHelpers.csproj
+++ b/tests/Agent/IntegrationTests/IntegrationTestHelpers/IntegrationTestHelpers.csproj
@@ -14,7 +14,6 @@
     <PackageReference Include="xunit" version="2.6.6" />
     <PackageReference Include="xunit.assert" Version="2.6.6" />
     <PackageReference Include="xunit.core" Version="2.6.6" />
-    <PackageReference Include="xunit.extensibility.core" Version="2.6.6" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Shared\Shared.csproj" />