--- a/tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/MultiFunctionApplicationHelpers.csproj
+++ b/tests/Agent/IntegrationTests/SharedApplications/Common/MultiFunctionApplicationHelpers/MultiFunctionApplicationHelpers.csproj
@@ -81,7 +81,6 @@
     <!-- MySql.Data .NET/Core references -->
     <PackageReference Include="MySql.Data" Version="6.10.7" Condition="'$(TargetFramework)' == 'net6.0'" />
     <!-- MySql.Data v8.0.33 is a breaking change for the agent and requires agent version 10.11.1 or greater -->
-    <PackageReference Include="MySql.Data" Version="8.3.0" Condition="'$(TargetFramework)' == 'net7.0'" />
     <PackageReference Include="MySql.Data" Version="8.3.0" Condition="'$(TargetFramework)' == 'net8.0'" />
 
     <!-- MongoDB.Driver .NET Framework references -->