--- a/tests/Agent/IntegrationTests/UnboundedIntegrationTests/UnboundedIntegrationTests.csproj
+++ b/tests/Agent/IntegrationTests/UnboundedIntegrationTests/UnboundedIntegrationTests.csproj
@@ -56,11 +56,6 @@
     <PackageReference Include="xunit" version="2.6.6" />
     <PackageReference Include="xunit.assert" Version="2.6.6" />
     <PackageReference Include="xunit.core" Version="2.6.6" />
-    <PackageReference Include="xunit.extensibility.core" Version="2.6.6" />
-    <PackageReference Include="xunit.runner.console" Version="2.6.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.6">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>